Peticiones de offer actualizadas

+++ src/main/java/com/campusdual/bfp/service/OfferService.java
@@ -10,6 +10,7 @@ import com.campusdual.bfp.model.dto.dtomapper.OfferMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.Date;
 import java.util.List;
 
 @Service
@@ -23,39 +24,44 @@ public class OfferService implements IOfferService {
 
     @Override
     public OfferDTO queryOffer(OfferDTO OfferDTO) {
-        Offer Offer = OfferMapper.INSTANCE.toEntity (OfferDTO);
+        Offer Offer = OfferMapper.INSTANCE.toEntity(OfferDTO);
         return OfferMapper.INSTANCE.toDTO(OfferDao.getReferenceById(Offer.getId()));
     }
+
     @Override
     public List<OfferDTO> queryAllOffers() {
         return OfferMapper.INSTANCE.toDTOList(OfferDao.findAll());
     }
+
     @Override
-    public int insertOffer(OfferDTO OfferDTO) {
+    public int insertOffer(OfferDTO request, String username) {
         User user = userDao.findByLogin(username);
         if (user == null) throw new RuntimeException("Usuario no encontrado");
-
         Offer offer = new Offer();
         offer.setTitle(request.getTitle());
         offer.setDescription(request.getDescription());
         offer.setActive(true);
         offer.setDate(new Date());
-        // Asigna el id de la empresa/usuario
         offer.setCompanyId(user.getId());
-
-        offerDao.save(offer);
-        return offer.getId();
+        Offer savedOffer = OfferDao.saveAndFlush(offer);
+        return savedOffer.getId();
     }
 
     @Override
-    public int updateOffer(OfferDTO OfferDTO) {
-        return insertOffer(OfferDTO);
+    public int updateOffer(OfferDTO request, String username) {
+        return insertOffer(request, username);
     }
+
     @Override
-    public int deleteOffer(OfferDTO OfferDTO) {
+    public int deleteOffer(OfferDTO OfferDTO, String username) {
+        User user = userDao.findByLogin(username);
+        if (user == null) throw new RuntimeException("Usuario no encontrado");
+        Offer existingOffer = OfferDao.findById(OfferDTO.getId())
+                .orElseThrow(() -> new RuntimeException("Oferta no encontrada con ID: " + OfferDTO.getId()));
+
         int id = OfferDTO.getId();
         Offer Offer = OfferMapper.INSTANCE.toEntity(OfferDTO);
         OfferDao.delete(Offer);
         return id;
     }
-    }
+}
